Extract Query to its own file
