+++ packages/alfa-media/src/media.ts
@@ -3,26 +3,15 @@ import { Device } from "@siteimprove/alfa-device";
 import { Equatable } from "@siteimprove/alfa-equatable";
 import { Iterable } from "@siteimprove/alfa-iterable";
 import { Serializable } from "@siteimprove/alfa-json";
-import { Option, None } from "@siteimprove/alfa-option";
 import { Parser } from "@siteimprove/alfa-parser";
 import { Predicate } from "@siteimprove/alfa-predicate";
 
 import * as json from "@siteimprove/alfa-json";
 
-import { Condition } from "./condition";
+import { Query } from "./query";
 import { Type } from "./type";
 
-const {
-  map,
-  either,
-  option,
-  pair,
-  left,
-  right,
-  delimited,
-  separatedList,
-  eof,
-} = Parser;
+const { map, option, left, delimited, separatedList, eof } = Parser;
 
 export namespace Media {
   export interface Queryable<T extends json.JSON = json.JSON>
@@ -30,19 +19,6 @@ export namespace Media {
       Serializable<T> {
     matches: Predicate<Device>;
   }
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#media-query-modifier
-   */
-  export enum Modifier {
-    Only = "only",
-    Not = "not",
-  }
-
-  const parseModifier = either(
-    map(Token.parseIdent("only"), () => Modifier.Only),
-    map(Token.parseIdent("not"), () => Modifier.Not)
-  );
 
   export enum Comparator {
     GreaterThan = ">",
@@ -53,125 +29,7 @@ export namespace Media {
 
   export const { of: type, isType } = Type;
 
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#media-query
-   */
-  export class Query implements Queryable<Query.JSON> {
-    public static of(
-      modifier: Option<Modifier>,
-      mediaType: Option<Type>,
-      condition: Option<Condition.Condition>
-    ): Query {
-      return new Query(modifier, mediaType, condition);
-    }
-
-    private readonly _modifier: Option<Modifier>;
-    private readonly _mediaType: Option<Type>;
-    private readonly _condition: Option<Condition.Condition>;
-
-    private constructor(
-      modifier: Option<Modifier>,
-      mediaType: Option<Type>,
-      condition: Option<Condition.Condition>
-    ) {
-      this._modifier = modifier;
-      this._mediaType = mediaType;
-      this._condition = condition;
-    }
-
-    public get modifier(): Option<Modifier> {
-      return this._modifier;
-    }
-
-    public get mediaType(): Option<Type> {
-      return this._mediaType;
-    }
-
-    public get condition(): Option<Condition.Condition> {
-      return this._condition;
-    }
-
-    public matches(device: Device): boolean {
-      const negated = this._modifier.some(
-        (modifier) => modifier === Modifier.Not
-      );
-      const type = this._mediaType.every((type) => type.matches(device));
-      const condition = this.condition.every((condition) =>
-        condition.matches(device)
-      );
-
-      return negated !== (type && condition);
-    }
-
-    public equals(value: unknown): value is this {
-      return (
-        value instanceof Query &&
-        value._modifier.equals(this._modifier) &&
-        value._mediaType.equals(this._mediaType) &&
-        value._condition.equals(this._condition)
-      );
-    }
-
-    public toJSON(): Query.JSON {
-      return {
-        modifier: this._modifier.getOr(null),
-        mediaType: this._mediaType.map((type) => type.toJSON()).getOr(null),
-        condition: this._condition
-          .map((condition) => condition.toJSON())
-          .getOr(null),
-        type: "query",
-      };
-    }
-
-    public toString(): string {
-      const modifier = this._modifier.getOr("");
-
-      const type = this._mediaType
-        .map((type) => (modifier === "" ? `${type}` : `${modifier} ${type}`))
-        .getOr("");
-
-      return this._condition
-        .map((condition) =>
-          type === "" ? `${condition}` : `${type} and ${condition}`
-        )
-        .getOr(type);
-    }
-  }
-
-  export namespace Query {
-    export interface JSON {
-      [key: string]: json.JSON;
-      modifier: string | null;
-      mediaType: Type.JSON | null;
-      condition: Condition.JSON | null;
-      type: "query";
-    }
-  }
-
-  /**
-   * @see https://drafts.csswg.org/mediaqueries/#typedef-media-query
-   */
-  const parseQuery = either(
-    map(Condition.parse, (condition) =>
-      Query.of(None, None, Option.of(condition))
-    ),
-    map(
-      pair(
-        pair(
-          option(delimited(option(Token.parseWhitespace), parseModifier)),
-          Type.parse
-        ),
-        option(
-          right(
-            delimited(option(Token.parseWhitespace), Token.parseIdent("and")),
-            Condition.parseWithoutOr
-          )
-        )
-      ),
-      ([[modifier, type], condition]) =>
-        Query.of(modifier, Option.of(type), condition)
-    )
-  );
+  export const { of: query, isQuery } = Query;
 
   /**
    * @see https://drafts.csswg.org/mediaqueries/#media-query-list
@@ -221,18 +79,22 @@ export namespace Media {
 
   export namespace List {
     export type JSON = Array<Query.JSON>;
+
+    export function isList(value: unknown): value is List {
+      return value instanceof List;
+    }
+
+    export const parse = left(
+      map(
+        separatedList(
+          Query.parse,
+          delimited(option(Token.parseWhitespace), Token.parseComma)
+        ),
+        (queries) => List.of(queries)
+      ),
+      eof((token) => `Unexpected token ${token}`)
+    );
   }
 
-  const parseList = left(
-    map(
-      separatedList(
-        parseQuery,
-        delimited(option(Token.parseWhitespace), Token.parseComma)
-      ),
-      (queries) => List.of(queries)
-    ),
-    eof((token) => `Unexpected token ${token}`)
-  );
-
-  export const parse = parseList;
+  export const parse = List.parse;
 }
